feat(nuxt-jsonld): customizing indentation implementation
to minify the generated json

BREAKING CHANGE: deprecate the static `mixin` and here comes `createMixin`, the mixin constructor, as the new public API

resolves #7
